raise exception of inversion performe with noise covariane
