+++ test_autogalaxy/conftest.py
@@ -106,6 +106,11 @@ def make_masked_imaging_7x7():
     return fixtures.make_masked_imaging_7x7()
 
 
+@pytest.fixture(name="masked_imaging_covariance_7x7")
+def make_masked_imaging_covariance_7x7():
+    return fixtures.make_masked_imaging_covariance_7x7()
+
+
 @pytest.fixture(name="masked_imaging_7x7_no_blur")
 def make_masked_imaging_7x7_no_blur():
     return fixtures.make_masked_imaging_7x7_no_blur()
